fix cc
